refactor([issue-744]): extract taskPromptService from taskSchedule.js
Move the default prompt catalog (DEFAULT_TASK_PROMPTS), prompt-version
machinery (PROMPT_VERSIONS, PREVIOUS_DEFAULT_PROMPTS,
REFERENCE_WATCH_AUDITED_VERSION), and the prompt getters
(getDefaultPrompt, getTaskPrompt, getStagePrompt, placeholder resolution)
out of the 3561-LOC taskSchedule.js into a new taskPromptService.js, one
concern per file.

The prompt data is moved verbatim — defaults are byte-identical, so the
PROMPT_VERSIONS / PREVIOUS_DEFAULT_PROMPTS auto-upgrade machinery is
preserved exactly (no default changes, no version bump). taskSchedule.js
still consumes those constants for loadSchedule()'s upgrade path, so it
imports them back and re-exports the public prompt API
(getDefaultPrompt/getTaskPrompt/getStagePrompt/PROMPT_VERSIONS/
REFERENCE_WATCH_AUDITED_VERSION) — existing importers are unaffected.

The getters' one runtime dependency on getTaskInterval is satisfied via a
lazy import('./taskSchedule.js') so the two modules don't form a static
circular import.

Refs #744.
